# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music API",
    "description": "A comprehensive REST API for managing artists, albums, songs, and playlists with full CRUD operations, authentication, and data validation",
    "version": "2.0.0",
    "contact": {
      "name": "Music API Support",
      "email": "support@musicapi.com"
    }
  },
  "servers": [
    {
      "url": "https://music-api-cse341-u669.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/albums": {
      "get": {
        "tags": ["Albums"],
        "summary": "Get all albums with filtering, sorting, and pagination",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "description": "Filter by genre",
            "schema": { "type": "string" }
          },
          {
            "name": "artist_id",
            "in": "query",
            "description": "Filter by artist ID", 
            "schema": { "type": "string" }
          },
          {
            "name": "year",
            "in": "query",
            "description": "Filter by release year",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Successfully retrieved albums" }
        }
      },
      "post": {
        "tags": ["Albums"],
        "summary": "Create a new album",
        "security": [{ "OAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AlbumInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Album created successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Authentication required" }
        }
      }
    },
    "/albums/{id}": {
      "get": {
        "tags": ["Albums"],
        "summary": "Get album by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Album found" },
          "404": { "description": "Album not found" }
        }
      },
      "put": {
        "tags": ["Albums"],
        "summary": "Update album",
        "security": [{ "OAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AlbumInput" }
            }
          }
        },
        "responses": {
          "204": { "description": "Album updated successfully" },
          "401": { "description": "Authentication required" }
        }
      },
      "delete": {
        "tags": ["Albums"],
        "summary": "Delete album",
        "security": [{ "OAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Album deleted successfully" },
          "401": { "description": "Authentication required" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlbumInput": {
        "type": "object",
        "required": ["title", "artist_id", "release_date", "genre", "track_count", "duration"],
        "properties": {
          "title": { "type": "string", "example": "Abbey Road" },
          "artist_id": { "type": "string", "example": "60f7c2b8d4f1c12345678901" },
          "release_date": { "type": "string", "format": "date", "example": "1969-09-26" },
          "genre": { "type": "string", "example": "Rock" },
          "track_count": { "type": "integer", "example": 17 },
          "duration": { "type": "integer", "example": 47 },
          "record_label": { "type": "string", "example": "Apple Records" },
          "cover_image_url": { "type": "string", "example": "https://example.com/abbey-road.jpg" }
        }
      },
...
    },
    "securitySchemes": {
      "OAuth": {
        "type": "oauth2",
        "description": "OAuth 2.0 authentication with Google",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/auth/google",
            "tokenUrl": "/auth/google/callback",
            "scopes": {
              "profile": "Access user profile",
              "email": "Access user email"
            }
          }
        }
      }
    }
  }
}